fixed pagination (on delete all)
