+++ js-2/db.json
@@ -1,176 +1,6 @@
 {
   "customers": [
     {
-      "_id": "5bf8de1a6e4b02b2a25ee3b0",
-      "name": "Fffjss",
-      "company": "aaa",
-      "address": "River Street 887",
-      "city": "Kirk, Hawaii, 1875",
-      "country": "Cyprus",
-      "email": "forbeshays@xurban.com",
-      "phone": "+1 (874) 417-3818"
-    },
-    {
-      "_id": "5bf8de1a33a4f6c1e01c9b27",
-      "name": "ffss",
-      "company": "aaaaaa",
-      "address": "Nautilus Avenue 561",
-      "city": "Fairhaven, District Of Columbia, 4743",
-      "country": "Mali",
-      "email": "muellerhensley@quailcom.com",
-      "phone": "+1 (892) 540-3422"
-    },
-    {
-      "_id": "5bf8de1a09b608036382ffe7",
-      "name": "ffsdasdsd",
-      "company": "jjjj",
-      "address": "Bleecker Street 997",
-      "city": "Wyano, Kansas, 481",
-      "country": "Kenya",
-      "email": "patebrooks@locazone.com",
-      "phone": "+1 (833) 428-2595"
-    },
-    {
-      "_id": "5bf8de1a5e45d8548f321c89",
-      "name": "Miriam Wilkins",
-      "company": "MONDICIL",
-      "address": "Front Street 936",
-      "city": "Motley, Wisconsin, 7449",
-      "country": "Namibia",
-      "email": "miriamwilkins@mondicil.com",
-      "phone": "+1 (823) 513-3803"
-    },
-    {
-      "_id": "5bf8de1a0c756f7f6528a173",
-      "name": "Faulkner Ramsey",
-      "company": "TEMORAK",
-      "address": "Stuyvesant Avenue 191",
-      "city": "Bend, Vermont, 4519",
-      "country": "Northern Mariana Islands",
-      "email": "faulknerramsey@temorak.com",
-      "phone": "+1 (840) 405-2537"
-    },
-    {
-      "_id": "5bf8de1a371af583953d1c5e",
-      "name": "Pearson Mcintosh",
-      "company": "VERTON",
-      "address": "Hegeman Avenue 840",
-      "city": "Loyalhanna, Wyoming, 9183",
-      "country": "Antigua and Barbuda",
-      "email": "pearsonmcintosh@verton.com",
-      "phone": "+1 (930) 545-3704"
-    },
-    {
-      "_id": "5bf8de1a274040ce73108c7c",
-      "name": "Aline Dale",
-      "company": "TOYLETRY",
-      "address": "Woodruff Avenue 919",
-      "city": "Sehili, Federated States Of Micronesia, 8707",
-      "country": "Reunion",
-      "email": "alinedale@toyletry.com",
-      "phone": "+1 (860) 529-2056"
-    },
-    {
-      "_id": "5bf8de1a47dff50d9bc9bacd",
-      "name": "Annette Burton",
-      "company": "STEELFAB",
-      "address": "Milton Street 707",
-      "city": "Jamestown, North Carolina, 520",
-      "country": "Mauritius",
-      "email": "annetteburton@steelfab.com",
-      "phone": "+1 (811) 409-3136"
-    },
-    {
-      "_id": "5bf8de1aa9521985cb695a23",
-      "name": "Felecia Dyer",
-      "company": "LUNCHPAD",
-      "address": "Homecrest Court 840",
-      "city": "Sultana, Alabama, 1203",
-      "country": "Sierra Leone",
-      "email": "feleciadyer@lunchpad.com",
-      "phone": "+1 (947) 571-2881"
-    },
-    {
-      "_id": "5bf8de1a04ad680c53ae3f73",
-      "name": "Blevins Zamora",
-      "company": "CUJO",
-      "address": "Thornton Street 330",
-      "city": "Thatcher, Mississippi, 1778",
-      "country": "Barbados",
-      "email": "blevinszamora@cujo.com",
-      "phone": "+1 (922) 502-2329"
-    },
-    {
-      "_id": "5bf8de1ad58f4e228a3a74e1",
-      "name": "Ingrid Holman",
-      "company": "COASH",
-      "address": "Keap Street 406",
-      "city": "Blandburg, Montana, 8261",
-      "country": "Cote D'Ivoire (Ivory Coast)",
-      "email": "ingridholman@coash.com",
-      "phone": "+1 (804) 502-2229"
-    },
-    {
-      "_id": "5bf8de1a70d07795770bc044",
-      "name": "Carpenter Church",
-      "company": "SCENTY",
-      "address": "Lancaster Avenue 771",
-      "city": "Leola, Maine, 1591",
-      "country": "Guam",
-      "email": "carpenterchurch@scenty.com",
-      "phone": "+1 (802) 538-2705"
-    },
-    {
-      "_id": "5bf8de1a85c6461c72473ea4",
-      "name": "Bertha George",
-      "company": "VIRVA",
-      "address": "Butler Place 252",
-      "city": "Gwynn, North Dakota, 9160",
-      "country": "Seychelles",
-      "email": "berthageorge@virva.com",
-      "phone": "+1 (919) 442-3187"
-    },
-    {
-      "_id": "5bf8de1a15c33fedf87a824c",
-      "name": "Francine Munoz",
-      "company": "PROFLEX",
-      "address": "Varet Street 342",
-      "city": "Stouchsburg, New Jersey, 9733",
-      "country": "Sao Tome and Principe",
-      "email": "francinemunoz@proflex.com",
-      "phone": "+1 (982) 405-2000"
-    },
-    {
-      "_id": "5bf8de1a04c3da1670cad8f1",
-      "name": "Celeste Mcfarland",
-      "company": "QUALITERN",
-      "address": "Hudson Avenue 221",
-      "city": "Dotsero, Idaho, 5168",
-      "country": "British Indian Ocean Territory",
-      "email": "celestemcfarland@qualitern.com",
-      "phone": "+1 (817) 408-3167"
-    },
-    {
-      "_id": "5bf8de1a51238b42d05f8492",
-      "name": "Kelley Poole",
-      "company": "CONFRENZY",
-      "address": "Cove Lane 267",
-      "city": "Bridgetown, Rhode Island, 8222",
-      "country": "Argentina",
-      "email": "kelleypoole@confrenzy.com",
-      "phone": "+1 (854) 409-2237"
-    },
-    {
-      "_id": "5bf8de1a1fe5deec0e72354f",
-      "name": "Bauer Bolton",
-      "company": "MEGALL",
-      "address": "Applegate Court 499",
-      "city": "Omar, Oklahoma, 8013",
-      "country": "Eritrea",
-      "email": "bauerbolton@megall.com",
-      "phone": "+1 (832) 532-2293"
-    },
-    {
       "_id": "5bf8de1abd7019acbdc7470c",
       "name": "Graciela Gaines",
       "company": "ORBAXTER",
@@ -321,19 +151,9 @@
       "phone": "+1 (807) 511-3949"
     },
     {
-      "_id": "5bf8de1adc668eeeba3ec410",
-      "name": "Bobbi Rutledge",
-      "company": "NETBOOK",
-      "address": "Lafayette Avenue 969",
-      "city": "Cascades, Virgin Islands, 9230",
-      "country": "Tanzania",
-      "email": "bobbirutledge@netbook.com",
-      "phone": "+1 (836) 588-3200"
-    },
-    {
       "_id": "5bf8de1a04bba78e9c2e699e",
       "name": "Rosemarie Little",
-      "company": "FLEXIGEN",
+      "company": "FL",
       "address": "Glendale Court 464",
       "city": "Taft, Northern Mariana Islands, 9704",
       "country": "Bolivia",
@@ -399,16 +219,6 @@
       "country": "Italy",
       "email": "lornamason@tropoli.com",
       "phone": "+1 (816) 590-3925"
-    },
-    {
-      "_id": "49998558840454028434010896123127054695",
-      "name": "555",
-      "company": "11111",
-      "address": "666",
-      "country": "1",
-      "city": "6667",
-      "email": "forbeshays@xurban.com",
-      "phone": "+1 (874) 417-3818"
     }
   ]
 }
